% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conq_barplot.R
\name{conq_barplot}
\alias{conq_barplot}
\title{Create a conq_barplot}
\usage{
conq_barplot(
  data,
  x = NULL,
  n = 100,
  ncols = NULL,
  fill = "#2f4f4f",
  auto_fill = FALSE,
  show_bins = FALSE,
  type = c("median", "mean"),
  ...
)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{x}{The variable that generates the quantile bins.}

\item{n}{The number of bins to use for binning the data}

\item{ncols}{The number of column to be used in the layout.}

\item{fill}{The color to use for the bars.}

\item{auto_fill}{If \code{TRUE}, use a different color for each category}

\item{show_bins}{If \code{TRUE}, show the bins on the x-axis.}

\item{type}{The type of statistic to use for the bars.}

\item{...}{Additional arguments to pass to the plot functions}
}
\value{
A \code{list} of ggplot objects.
}
\description{
\code{\link[=conq_barplot]{conq_barplot()}} shows the median or mean for each \code{\link[=qbin]{qbin()}},
with on the x-axis the values of the conditional variable.
}
\examples{
data("diamonds", package="ggplot2")

table_plot(diamonds[c(1:4, 7)], "carat")

qbin_barplot(iris, "Sepal.Length", n = 25)
table_plot(iris, "Sepal.Length", 25)

if (require(palmerpenguins)) {
  table_plot(penguins[1:7], "body_mass_g", 25)
}


}
\seealso{
Other conditional quantile plotting functions: 
\code{\link{conq_boxplot}()},
\code{\link{funq_plot}()}
}
\concept{conditional quantile plotting functions}
