% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conq_plot.R
\name{conq_plot}
\alias{conq_plot}
\title{conditional quantile boxplot}
\usage{
conq_plot(
  data,
  x = NULL,
  n = 100,
  color = "#002f2f",
  fill = "#2f4f4f",
  auto_fill = FALSE,
  ncols = NULL,
  xmarker = NULL,
  qmarker = NULL,
  add_rug = FALSE,
  xlim = NULL,
  connect = FALSE,
  ...
)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{x}{\code{character} name of variable that should be plotted in the x-axis.}

\item{n}{The number of bins to use for binning the data}

\item{color}{The color to use for the line charts}

\item{fill}{The fill color to use for the areas}

\item{auto_fill}{If \code{TRUE}, use a different color for each category}

\item{ncols}{The number of column to be used in the layout}

\item{xmarker}{\code{numeric}, the x marker.}

\item{qmarker}{\code{numeric}, the quantile marker to use that is translated in a x value.}

\item{add_rug}{if \code{TRUE} a rug is added to the plot}

\item{xlim}{\code{numeric}, the limits of the x-axis.}

\item{connect}{if \code{TRUE} subsequent medians are connected.}

\item{...}{Additional arguments to pass to the plot functions}
}
\description{
Create a conditional quantile boxplot, show the boxplots of the quantile bins, with
on the x-axis the values of the conditional variable.
}
\details{
\code{conq_plot} is the same function as \code{\link[=funq_plot]{funq_plot()}} but with a different default,
namely \code{connect = FALSE}. \code{funq_plot} highlights the functional relationship between
x and the y-variables, but connecting the medians of the quantile bins.
\code{\link[=qbin_boxplot]{qbin_boxplot()}} shows the boxplots of the quantile bins on a quantile scale.
}
\examples{
conq_plot(iris, x = "Petal.Length", n = 25)
}
