% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependence_plot.R
\name{dependence_plot}
\alias{dependence_plot}
\title{Create a tableplot from a dataset}
\usage{
dependence_plot(data, sort_variable = NULL, n = 100, ncols = NULL, ...)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{sort_variable}{The variable to sort the data by}

\item{n}{The number of bins to use for binning the data}

\item{ncols}{The number of column to be used in the layout}

\item{...}{Additional arguments to pass to the plot functions}
}
\value{
A ggplot object
}
\description{
Create a table plot from a dataset/frame
}
\examples{
dependence_plot(iris, "Petal.Length", 20)

dependence_plot(diamonds, "price")
dependence_plot(diamonds, "carat")

if (require(palmerpenguins)){
  dependence_plot(penguins[1:7], "body_mass_g", 25)
  quantile_dependence_plot(penguins[1:7], c("body_mass_g"), 25)
}

}
