% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funq_plot.R
\name{funq_plot}
\alias{funq_plot}
\title{Plot variables as a function of another variable}
\usage{
funq_plot(data, x = NULL, n = 100, ncols = NULL, ...)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{x}{\code{character} name of variable that should be plotted in the x-axis.}

\item{n}{The number of bins to use for binning the data}

\item{ncols}{The number of column to be used in the layout}

\item{...}{Additional arguments to pass to the plot functions}
}
\value{
A ggplot object with the plots
}
\description{
Plot the median and interquartile range of numerical variables and frequency of
other variables as a function the \code{x} variable.
}
\details{
The statistics are derived by binning the \code{x} variable into \code{n} bins or quantiles and
calculating the statistics for each bin, hence the name \code{funq_plot}.
}
\examples{
funq_plot(iris, "Petal.Length", 20)

funq_plot(diamonds, "price")
funq_plot(diamonds, "carat")

if (require(palmerpenguins)){
  funq_plot(penguins[1:7], "body_mass_g", 25)
  quantile_dependence_plot(penguins[1:7], c("body_mass_g"), 25)
}

}
