% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtableplot.R
\name{ggtableplot}
\alias{ggtableplot}
\title{Create a tableplot from a dataset}
\usage{
ggtableplot(
  data,
  sort_variable = names(data)[1],
  n = 100,
  ncols = ncol(data),
  ...
)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{sort_variable}{The variable to sort the data by}

\item{n}{The number of bins to use for binning the data}

\item{ncols}{The number of column to be used in the layout}

\item{...}{Additional arguments to pass to the plot functions}
}
\value{
A ggplot object
}
\description{
Create a table plot from a dataset/frame
}
\examples{
data("diamonds", package="ggplot2")

ggtableplot(diamonds, "carat", nco=NULL)

ggtableplot(iris, "Sepal.Length", 25)

}
