% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdp.R
\name{partial_dependence_plot}
\alias{partial_dependence_plot}
\title{Create a tableplot from a dataset}
\usage{
partial_dependence_plot(
  data,
  sort_variable = names(data)[1],
  n = 100,
  ncols = ncol(data),
  ...
)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{sort_variable}{The variable to sort the data by}

\item{n}{The number of bins to use for binning the data}

\item{ncols}{The number of column to be used in the layout}

\item{...}{Additional arguments to pass to the plot functions}
}
\value{
A ggplot object
}
\description{
Create a table plot from a dataset/frame
}
\examples{
data("diamonds", package="ggplot2")

table_plot(diamonds, "carat")

table_plot(diamonds[c(1:4, 7)], "carat")

table_plot(iris, "Sepal.Length", 50)


if (require(palmerpenguins)) {
  table_plot(penguins[1:7], "body_mass_g", 25)
}
}
