% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile_dep.R
\name{percentile_dep_plot}
\alias{percentile_dep_plot}
\title{Create a percentile dependency plot}
\usage{
percentile_dep_plot(
  data,
  sort_variable = names(data)[1],
  n = 100,
  ncols = ncol(data),
  ...
)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{sort_variable}{The variable to sort the data by}

\item{n}{The number of bins to use for binning the data}

\item{ncols}{The number of column to be used in the layout}

\item{...}{Additional arguments to pass to the plot functions}
}
\value{
A ggplot object
}
\description{
Create a table plot from a dataset/frame
}
\examples{
data("diamonds", package="ggplot2")

percentile_dep_plot(diamonds[c(1:3,7)], "carat")

}
