% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbin_barplot.R, R/table_plot.R
\name{qbin_barplot}
\alias{qbin_barplot}
\alias{table_plot}
\title{Create a qbin_barplot}
\usage{
qbin_barplot(
  data,
  x = NULL,
  n = 100,
  ncols = NULL,
  fill = "#2f4f4f",
  type = c("median", "mean"),
  ...
)

table_plot(data, x = NULL, n = 100, ncols = ncol(data), fill = "#555555", ...)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{x}{The variable that generates the quantile bins.}

\item{n}{The number of bins to use for binning the data}

\item{ncols}{The number of column to be used in the layout.}

\item{fill}{The color to use for the bars.}

\item{type}{The type of statistic to use for the bars.}

\item{...}{Additional arguments to pass to the plot functions}
}
\value{
A \code{list} of ggplot objects.
}
\description{
\code{\link[=qbin_barplot]{qbin_barplot()}} shows the median or mean for each qbin, thereby focusing on
the expected value per bin.
}
\details{
The \code{table_plot} calls \code{qbin_barplot} with \code{ncols} set to \code{ncol(data)}
}
\examples{
data("diamonds", package="ggplot2")

table_plot(diamonds[c(1:4, 7)], "carat")

qbin_barplot(iris, "Sepal.Length", n = 25)
table_plot(iris, "Sepal.Length", 25)

if (require(palmerpenguins)) {
  table_plot(penguins[1:7], "body_mass_g", 25)
}


}
\seealso{
Other qbin plotting functions: 
\code{\link{qbin_boxplot}()},
\code{\link{qbin_heatmap}()}
}
\concept{qbin plotting functions}
