% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbin_boxplot.R
\name{qbin_boxplot}
\alias{qbin_boxplot}
\title{Create a quantile dependence plot}
\usage{
qbin_boxplot(data, sort_variable = NULL, n = 100, ncols = NULL, ...)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{sort_variable}{The variable to sort the data by}

\item{n}{The number of bins to use for binning the data}

\item{ncols}{The number of column to be used in the layout}

\item{...}{Additional arguments to pass to the plot functions}
}
\value{
A ggplot object
}
\description{
Create a quantile dependence plot
}
\examples{
data("diamonds", package="ggplot2")

qbin_boxplot(iris)
qbin_boxplot(diamonds[1:7], "carat")

}
