% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbin_heatmap.R
\name{qbin_heatmap}
\alias{qbin_heatmap}
\title{Create a qbin_heatmap from a dataset}
\usage{
qbin_heatmap(
  data,
  x = NULL,
  n = 100,
  ncols = NULL,
  auto_fill = FALSE,
  fill = "#2f4f4f",
  ...
)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{x}{The variable that generates the quantile bins.}

\item{n}{The number of bins to use for binning the data}

\item{ncols}{The number of column to be used in the layout.}

\item{auto_fill}{If \code{TRUE}, use a different color for each category.}

\item{fill}{The color to use for the heatmap.}

\item{...}{Additional arguments to pass to the plot functions}
}
\value{
A ggplot object
}
\description{
Create qbin_heatmap from a dataset/frame. Shows the distribution
of variables for each quantile bin of \code{x}.
}
\examples{

qbin_heatmap(iris, "Sepal.Length", n = 25)

data("diamonds", package="ggplot2")

qbin_heatmap(
  diamonds[c(1:4, 7)],
  "carat",
  color = "#333",
  auto_fill = TRUE
)




}
