% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbinplots-package.R
\docType{package}
\name{qbinplots-package}
\alias{qbinplots}
\alias{qbinplots-package}
\title{qbinplots}
\description{
This package creates plots using quantile binning.
}
\section{plot types}{

\itemize{
\item \code{\link[=qbin_boxplot]{qbin_boxplot()}} creates quantile bin boxplots, highlighting the change in median between qbins.
\item \code{\link[=qbin_barplot]{qbin_barplot()}} creates quantile bin barplots highlighting the size of means of qbins.
\item \code{\link[=qbin_heatmap]{qbin_heatmap()}} creates quantile bin heatmaps highlighting the distribution within the qbins.
\item \code{\link[=funq_plot]{funq_plot()}} creates a functional view of the data, plotting the median and interquartile range of numerical variables and level frequency of the other variables as a function of the \code{x} variable using quantile bins.
}
}

\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com} (\href{https://orcid.org/0000-0002-6580-4718}{ORCID})

}
