% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_dependence_plot.R
\name{quantile_dependence_plot}
\alias{quantile_dependence_plot}
\title{Create a quantile dependence plot}
\usage{
quantile_dependence_plot(
  data,
  sort_variable = names(data)[1],
  n = 100,
  ncols = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{sort_variable}{The variable to sort the data by}

\item{n}{The number of bins to use for binning the data}

\item{ncols}{The number of column to be used in the layout}

\item{...}{Additional arguments to pass to the plot functions}
}
\value{
A ggplot object
}
\description{
Create a quantile dependence plot
}
\examples{
data("diamonds", package="ggplot2")

quantile_dependence_plot(diamonds, "carat")

}
