% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funq_plot.R
\name{funq_plot}
\alias{funq_plot}
\title{Plot variables as a function of another variable}
\usage{
funq_plot(
  data,
  x = NULL,
  n = 100,
  color = "#002f2f",
  fill = "#2f4f4f",
  auto_fill = FALSE,
  ncols = NULL,
  xmarker = NULL,
  qmarker = NULL,
  show_bins = FALSE,
  xlim = NULL,
  connect = TRUE,
  ...
)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{x}{\code{character} name of variable that should be plotted in the x-axis.}

\item{n}{The number of bins to use for binning the data}

\item{color}{The color to use for the line charts}

\item{fill}{The fill color to use for the areas}

\item{auto_fill}{If \code{TRUE}, use a different color for each category}

\item{ncols}{The number of column to be used in the layout}

\item{xmarker}{\code{numeric}, the x marker.}

\item{qmarker}{\code{numeric}, the quantile marker to use that is translated in a x value.}

\item{show_bins}{if \code{TRUE} a rug is added to the plot}

\item{xlim}{\code{numeric}, the limits of the x-axis.}

\item{connect}{if \code{TRUE} subsequent medians are connected.}

\item{...}{Additional arguments to pass to the plot functions}
}
\value{
A ggplot object with the plots
}
\description{
Plot the median and interquartile range of numerical variables and level frequency
of the other variables as a function the \code{x} variable using quantile bins,
hence the name \code{funq_plot}.
}
\details{
It \code{qbin}s the \code{x} variable and plots the medians of the qbins vs the other variables, thereby
creating a functional view of \code{x} to the rest of the data,
calculating the statistics for each bin, hence the name \code{funq_plot}.
}
\examples{
funq_plot(iris, "Sepal.Length", n = 25, auto_fill = TRUE, xmarker=5.5)

funq_plot(diamonds[1:7], "carat", xlim=c(0,2))

if (require(palmerpenguins)){
  funq_plot(penguins[1:7], "body_mass_g", n = 25, auto_fill = TRUE, xmarker=4650)
}

}
\seealso{
Other conditional quantile plotting functions: 
\code{\link{conq_barplot}()},
\code{\link{conq_boxplot}()}
}
\concept{conditional quantile plotting functions}
