% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbin.R
\name{qbin}
\alias{qbin}
\title{Bin a data.frame into quantile bins}
\usage{
qbin(data, x = NULL, n = 100, min_bin_size = 5, ...)
}
\arguments{
\item{data}{a \code{data.frame} to be binned}

\item{x}{\code{character} variable name used for the quantile binning}

\item{n}{\code{integer} number of quantile bins.}

\item{min_bin_size}{\code{integer} minimum number of rows/data points that should be
in a quantile bin.}

\item{...}{reserved for future use}
}
\description{
Bins a data.frame into quantile bins for variable \code{x} in \code{data}.
}
\details{
Each numeric variable in the data.frame is binned into \code{n} quantile bins, for
which the \code{\link[=fivenum]{fivenum()}} and \code{\link[=mean]{mean()}} is calculated.

Each categorical variable is binned into \code{n} quantile bins, for which the
level frequency is calculated.
}
