% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbin_barplot.R, R/table_plot.R
\name{qbin_barplot}
\alias{qbin_barplot}
\alias{table_plot}
\title{Create a qbin_barplot from a dataset}
\usage{
qbin_barplot(data, x = NULL, n = 100, ncols = NULL, ...)

table_plot(data, x = NULL, n = 100, ncols = ncol(data), ...)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{x}{The variable that generates the quantile bins.}

\item{n}{The number of bins to use for binning the data}

\item{ncols}{The number of column to be used in the layout}

\item{...}{Additional arguments to pass to the plot functions}
}
\value{
A ggplot object
}
\description{
Create qbin_barplot/table_plot from a dataset/frame. Shows the distribution
of variables for each quantile bin of \code{x}.
}
\details{
The \code{table_plot} calls \code{qbin_barplot} with \code{ncols} set to \code{ncol(data)}
}
\examples{
data("diamonds", package="ggplot2")

table_plot(diamonds, "carat")

table_plot(diamonds[c(1:4, 7)], "carat")

table_plot(iris, "Sepal.Length", 50)


if (require(palmerpenguins)) {
  table_plot(penguins[1:7], "body_mass_g", 25)
}
}
