% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbin_boxplot.R
\name{qbin_boxplot}
\alias{qbin_boxplot}
\title{Create a quantile dependence plot}
\usage{
qbin_boxplot(
  data,
  sort_variable = NULL,
  n = 100,
  ncols = NULL,
  color = "darkblue",
  fill = "#555555",
  auto_fill = FALSE,
  qmarker = NULL,
  xmarker = NULL,
  ...
)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{sort_variable}{The variable to sort the data by}

\item{n}{The number of bins to use for binning the data}

\item{ncols}{The number of column to be used in the layout}

\item{color}{The color to use for the lines}

\item{fill}{The color to use for the bars}

\item{auto_fill}{If \code{TRUE}, use a different color for each category}

\item{qmarker}{\code{numeric}, the quantile marker to use.}

\item{xmarker}{\code{numeric} the x marker, i.e. the value for x that is translated into a q value.}

\item{...}{Additional arguments to pass to the plot functions}
}
\value{
A ggplot object
}
\description{
Create a quantile dependence plot
}
\examples{
data("diamonds", package="ggplot2")

qbin_boxplot(iris, auto_fill = TRUE)

qbin_boxplot(diamonds[1:7], "carat", auto_fill = TRUE, color = "black")
qbin_boxplot(diamonds[1:7], "price", auto_fill = TRUE, color = "black")




}
