% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qbin_heatmap.R
\name{qbin_heatmap}
\alias{qbin_heatmap}
\title{Create a qbin_heatmap from a dataset}
\usage{
qbin_heatmap(
  data,
  x = NULL,
  n = 100,
  bins = n,
  type = c("gradient", "size"),
  ncols = NULL,
  auto_fill = FALSE,
  fill = "#2f4f4f",
  low = "#eeeeee",
  high = "#2f4f4f",
  ...
)
}
\arguments{
\item{data}{A data.frame or data.table}

\item{x}{The variable that generates the quantile bins.}

\item{n}{The number of bins to use for binning the data, is overruled by \code{bins}}

\item{bins}{\code{integer} vector with the number of bins to use for the x and y axis.}

\item{type}{The type of heatmap to use. Either "gradient" or "size".}

\item{ncols}{The number of column to be used in the layout.}

\item{auto_fill}{If \code{TRUE}, use a different color for each category.}

\item{fill}{The color used for categorical variables.}

\item{low}{The color used for low values in the heatmap.}

\item{high}{The color used for high values in the heatmap.}

\item{...}{Additional arguments to pass to the plot functions}
}
\value{
A ggplot object
}
\description{
\code{qbin_heatmap} from a dataset/frame. Shows the distribution
of variables for each quantile bin of \code{x}. It is a complement to
\code{\link[=qbin_boxplot]{qbin_boxplot()}}, focussing on the distribution per \code{\link[=qbin]{qbin()}}.
}
\examples{

qbin_heatmap(
  iris,
  "Sepal.Length",
  n = 25,
  auto_fill = TRUE
)

qbin_heatmap(
  iris,
  "Sepal.Length",
  n = 25,
  auto_fill = TRUE,
  type = "size"
)


data("diamonds", package="ggplot2")

qbin_heatmap(
  diamonds[c(1,7:9)],
  "carat",
)




}
\concept{qbin plotting functions}
